requirements.txt
fastapi>=0.100.0
uvicorn==0.23.2
python-dotenv==1.0.0
langchain>=0.0.310
langchain-openai>=0.0.2  
openai>=0.27.8  
faiss-cpu==1.7.4
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
pydantic>=2.0.0  